win32: another attempt at a UCRT hack
The problem with UCRT seems to be that if a process has been
started with a non-NULL environment block passed to CreateProcess()
any subsequent call to spawnve() with a non-NULL environment pointer
fails.

Commit 5b48ca53b (win32: pass NULL to spawnve, not environ) fixed
the problem in busybox-w32 for those cases where a NULL environment
pointer was sufficient.  It didn't handle the case where the shell
passes a modified environment to its child.

All calls to spawnve() in the shell occur in a process which will
terminate whether or not the call succeeds.  It therefore doesn't
matter if we mess with the environment of this process such that
spawnve() can be passed a NULL environment pointer.  (I think.)